requirements.txt
torch==2.3.0
torchaudio==2.3.0
wandb==0.16.3
numpy==1.26.4
PyYAML==6.0.1
tqdm==4.66.2
hydra-core==1.3
